+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Make `Down.open` pass encoding from content type charset to `Down::ChunkedIO` (@janko-m)
+
 * Add `:encoding` option to `Down::ChunkedIO.new` for specifying the encoding of returned content (@janko-m)
 
 * Add HTTP.rb backend as an alternative to Net::HTTP (@janko-m)